auth-service/requirements.txt
flask==2.0.1
werkzeug==2.0.3
authlib==1.0.1
requests==2.28.2
pymongo==4.3.3
python-dotenv==0.21.0